move api creation to a page handle a few other states, remove raw error
