+++ apps/web/app/(authenticated)/(app)/app/[apiId]/keys/new/page.tsx
@@ -1,0 +1,25 @@
+import { getTenantId } from "@/lib/auth";
+import { db, schema, eq } from "@unkey/db";
+import { redirect } from "next/navigation";
+import { CreateKey } from "./CreateKey";
+
+export default async function ApiPage(props: { params: { apiId: string } }) {
+  const tenantId = getTenantId();
+
+  const api = await db.query.apis.findFirst({
+    where: eq(schema.apis.id, props.params.apiId),
+    with: {
+      workspace: true,
+      keys: true,
+    },
+  });
+  if (!api || api.workspace.tenantId !== tenantId) {
+    return redirect("/onboarding");
+  }
+
+  return (
+    <div className="h-full">
+      <CreateKey apiId={api.id} />
+    </div>
+  );
+}
