+++ apps/web/app/(authenticated)/(app)/app/apis/page.tsx
@@ -15,6 +15,7 @@ import {
 } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 
+
 export default async function TenantOverviewPage() {
   const tenantId = getTenantId();
   const workspace = await db.query.workspaces.findFirst({
